feat(elemento-rup): filtro por sexo en requeridos
