+++ src/app/modules/rup/components/elementos/deprecated/informeEpicrisis.component.ts
@@ -35,19 +35,6 @@ export class InformeEpicrisisComponent extends RUPComponent implements OnInit {
         });
     }
 
-    get requeridos() {
-        const requeridos = [];
-        for (let i = 0; i < this.registro.registros.length; i++) {
-            const concepto = this.registro.registros[i].concepto;
-            const requerido = this.elementoRUP.requeridos.find(r => r.concepto.conceptId === concepto.conceptId);
-            if (requerido) {
-                requeridos.push(requerido);
-            }
-        }
-        return requeridos;
-    }
-
-
     accordionSeleccionado(i, concepto: any) {
         if (this.accordionActive === i) {
             this.accordionActive = -1;
