+++ src/app/modules/rup/interfaces/elementoRUP.interface.ts
@@ -38,20 +38,23 @@ export interface IElementoRUP {
     conceptosBuscar: ISnomedConcept[];
     // Elementos RUP requeridos para la ejecución.
     // Por ejemplo, en "Control de Niño sano" es obligatorio ejecutar "Toma de peso"
-    requeridos: [{
-        elementoRUP: IElementoRUP,
-        concepto: ISnomedConcept,
-        // Indica estilos para la instancia del elementoRUP
-        style: {
-            columns: Number,
-            cssClass: String
-        },
-        // Indica parámetros para la instancia del elementoRUP en formato {key: value}
-        params: any
-    }];
+    requeridos: IElementoRUPRequeridos[];
     // Elementos RUP más frecuentes para la ejecución.
     // Por ejemplo, en "Consulta de medicina general" se puede sugerir ejecutar "Signos vitales"
     frecuentes: ISnomedConcept[];
 
     inactiveAt?: Date;
+}
+
+export interface IElementoRUPRequeridos {
+    elementoRUP: IElementoRUP;
+    concepto: ISnomedConcept;
+    // Indica estilos para la instancia del elementoRUP
+    style: {
+        columns: Number,
+        cssClass: String
+    };
+    // Indica parámetros para la instancia del elementoRUP en formato {key: value}
+    params: { [key: string]: any };
+    sexo: string;
 }
