+++ src/app/modules/rup/interfaces/prestacion.registro.interface.ts
@@ -1,7 +1,8 @@
-import { IElementoRUP } from './elementoRUP.interface';
+import { IElementoRUP, IElementoRUPRequeridos } from './elementoRUP.interface';
 import { ISnomedConcept } from './snomed-concept.interface';
 import { ObjectID } from 'bson';
 import { IPaciente } from '../../../core/mpi/interfaces/IPaciente';
+import { IPrestacion } from './prestacion.interface';
 
 export class IRegistroPrivacy {
     scope: string;
@@ -48,7 +49,7 @@ export class IPrestacionRegistro {
     // Virtuales 🤷
     paciente: IPaciente;
 
-    constructor(elementoRUP: IElementoRUP, snomedConcept: ISnomedConcept) {
+    constructor(elementoRUP: IElementoRUP, snomedConcept: ISnomedConcept, prestacion?: IPrestacion) {
         this.id = (new ObjectID()).toString();
         this.elementoRUP = elementoRUP ? elementoRUP.id : null;
         this.nombre = snomedConcept.term;
@@ -63,9 +64,18 @@ export class IPrestacionRegistro {
         this.noIndex = false;
         if (elementoRUP && elementoRUP.requeridos) {
             elementoRUP.requeridos.forEach((item) => {
-                this.registros.push(new IPrestacionRegistro(item.elementoRUP, item.concepto));
+                const canAdd = this.checkSexRule(prestacion, item);
+                if (canAdd) {
+                    this.registros.push(new IPrestacionRegistro(item.elementoRUP, item.concepto, prestacion));
+                }
             });
         }
 
+    }
+
+    private checkSexRule(prestacion: IPrestacion, requerido: IElementoRUPRequeridos) {
+        const sexo = prestacion && prestacion.paciente && prestacion.paciente.sexo;
+        const sexoFilter = requerido && requerido.sexo;
+        return !sexo || !sexoFilter || sexo === sexoFilter;
     }
 }
